tech-recon/requirements_web.txt
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
python-dotenv>=1.1.1
